# services/api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    name: Stanley Emeh
  description: API Gateway for the EucastanPay Microservices Platform.
  title: EucastanPay API Gateway
  version: "1.0"
paths:
...
  /admin/audits:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: List AuditReads
      tags:
      - Audit
  /admin/audits/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Audit Read ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Get AuditRead by ID
      tags:
      - Audit
  /admin/audits/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Event type
        in: query
        name: event_type
        type: string
      - description: Correlation ID
        in: query
        name: correlation_id
        type: string
      - description: Reference
        in: query
        name: reference
        type: string
      - description: Minimum amount
        in: query
        name: min_amount
        type: integer
      - description: Maximum amount
        in: query
        name: max_amount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Search AuditRead
      tags:
      - Audit
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"